+++ src/devices/GoveeRGBICLight.ts
@@ -1,18 +1,15 @@
 import {DeviceConfig} from '../core/structures/devices/DeviceConfig';
-import {SolidColor} from './states/SolidColor';
-import {MusicMode} from './states/MusicMode';
 import {GoveeLight} from './GoveeLight';
 import {DeviceFactory} from './DeviceFactory';
+import {ColorSegments} from './states/ColorSegments';
 
 
 @DeviceFactory.register(
   'H611A',
 )
 export class GoveeRGBICLight
-  extends MusicMode(
-    SolidColor(
-      GoveeLight,
-    ),
+  extends ColorSegments(
+    GoveeLight,
   ) {
 
   constructor(
