+++ src/platform/PlatformModule.ts
@@ -19,9 +19,10 @@ import {HumidifierService} from './accessories/services/HumidifierService';
 import {InformationService} from './accessories/services/InformationService';
 import {LightService} from './accessories/services/LightService';
 import {PurifierService} from './accessories/services/PurifierService';
-import {ColorTemperature} from './accessories/services/lightCharacteristics/ColorTemperature';
-import {Hue} from './accessories/services/lightCharacteristics/Hue';
-import {Saturation} from './accessories/services/lightCharacteristics/Saturation';
+import {ColorTemperature} from './accessories/services/lights/ColorTemperature';
+import {Hue} from './accessories/services/lights/Hue';
+import {Saturation} from './accessories/services/lights/Saturation';
+import {LightColorSegments} from './accessories/services/lights/LightColorSegments';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const ACCESSORY_SERVICES = [
@@ -32,6 +33,7 @@ const ACCESSORY_SERVICES = [
   ColorTemperature,
   Hue,
   Saturation,
+  LightColorSegments,
 ];
 
 export interface GoveeCredentials {
