+++ src/platform/accessories/services/lights/ColorTemperature.ts
