+++ src/util/const.ts
@@ -21,3 +21,5 @@ export const HOMEBRIDGE_API = 'Homebridge_Api';
 
 export const COMMAND_IDENTIFIER = 0x33;
 export const REPORT_IDENTIFIER = 0xaa;
+
+export const SEGMENT_COUNT = 15;
